backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
